+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - prepare-commit-msg git hook
 - enable package discovery for Laravel 5.5
 - added post-checkout git hook #9
-- add composer command #10
+- add install dependencies command #10
 
 ### Deprecated
 - Nothing